backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
breezy==3.0.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.3
configobj==5.0.6
cryptography==36.0.1
Cython==0.29.25
dbus-python==1.2.16
dulwich==0.19.15
fastapi==0.73.0
fastimport==0.9.8
h11==0.13.0
httptools==0.3.0
idna==2.8
outcome==1.1.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.3.1
PyGObject==3.36.0
pyOpenSSL==22.0.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
selenium==4.1.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.17.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wsproto==1.0.0
